不具合個所の修正 close #28
